Fixes after code review
