--- lib/tweet_item_model.dart
@@ -1,8 +1,0 @@
-class TweetItemModel {
-  String tweet;
-  String username;
-  String time;
-  String twitterHandle;
-
-  TweetItemModel(this.tweet, this.username, this.time, this.twitterHandle);
-}
